#332: Rename base package from sb3 to sb4
