+++ sb4/src/main/java/de/aikiit/prototype4/login/LoginTenantAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package de.aikiit.prototype3.login;
+package de.aikiit.prototype4.login;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
