+++ sb4/src/main/java/de/aikiit/prototype4/seeding/BootstrapDataCreator.java
@@ -1,9 +1,9 @@
-package de.aikiit.prototype3.seeding;
+package de.aikiit.prototype4.seeding;
 
 
-import de.aikiit.prototype3.tenant.TenantRepository;
-import de.aikiit.prototype3.user.RoleRepository;
-import de.aikiit.prototype3.user.UserRepository;
+import de.aikiit.prototype4.tenant.TenantRepository;
+import de.aikiit.prototype4.user.RoleRepository;
+import de.aikiit.prototype4.user.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
